packages/embcli-voyage/pyproject.toml
[project]
name = "embcli-voyage"
version = "0.1.1-dev1"
description = "voyage plugin for embcli"
readme = "README.md"
authors = [{ name = "Tomoko Uchida", email = "tomoko.uchida.1111@gmail.com" }]
license = { text = "Apache-2.0" }
keywords = ["cli", "llm", "nlp", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Terminals",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = ["embcli-core>=0.1.0", "pillow>=11.2.1", "voyageai>=0.3.2"]

[project.urls]
"Repository" = "https://github.com/mocobeta/embcli"
"Homepage" = "https://embcli.mocobeta.dev/"

[tool.uv.sources]
embcli-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."embcli"]
voyage = "embcli_voyage.voyage"
voyage-multimodal = "embcli_voyage.voyage_multimodal"
